fix(modelservice): 输入/输出成本字段支持小数
- ProviderConfig.costPerInputToken / costPerOutputToken：int → double
- LlmConfigProperties / CreateModelServiceRequest / UpdateModelServiceRequest /
  ModelServiceResponse 字段类型同步改 Double
- ModelServiceController.buildMetadata + 新增 getDoubleMetadata 处理 metadata_json
  反序列化（Jackson 兼容 Number / Integer / String 三种值类型）
- estimateCost 返回 double，内部用 1_000_000.0 浮点除法
- 前端 ModelServiceManager 成本 input 加 type="number" step="0.01" min="0"

+++ src/main/java/com/lifepilot/llm/config/LlmConfigProperties.java
@@ -99,8 +99,8 @@ public class LlmConfigProperties {
         private List<String> scenes = List.of();
         private Set<ProviderCapability> capabilities = Set.of(ProviderCapability.CHAT);
         private boolean enabled = true;
-        private int costPerInputToken = 0;
-        private int costPerOutputToken = 0;
+        private double costPerInputToken = 0;
+        private double costPerOutputToken = 0;
         private int maxContextWindow = 4096;
         private Integer embeddingDimension;
         private boolean supportsStreaming = false;
